Move helper function to separate files
