+++ src/utils/countChars.ts
@@ -1,0 +1,3 @@
+export default function countChars(event: React.ChangeEvent<HTMLInputElement>) {
+  return event.target.value.length <= 255;
+}
