+++ src/utils/countOneWord.ts
@@ -1,0 +1,5 @@
+export default function countOneWord(
+  event: React.ChangeEvent<HTMLInputElement>,
+) {
+  return event.target.value.split(' ').length === 1;
+}
